+++ src/compute-engine/symbolic/derivative.ts
@@ -128,7 +128,7 @@ const DERIVATIVES_TABLE = {
     ],
   ],
   // d/dx ln(Γ(x)) = ψ(x) (digamma function)
-  LogGamma: ['Digamma', '_'],
+  GammaLn: ['Digamma', '_'],
   // d/dx ψ(x) = ψ₁(x) (trigamma function)
   // https://en.wikipedia.org/wiki/Trigamma_function
   Digamma: ['Trigamma', '_'],